pyproject.toml
[project]
name = "astralconnector"
version = "0.1.0"
description = ""
authors = [
    {name = "Cosmic Onyx",email = "c0sm1c0nyx.225@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2,<6.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "djoser (==2.2.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "channels-redis (>=4.2.1,<5.0.0)",
    "uvicorn[standard] (>=0.34.2,<0.35.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "django-ratelimit (>=4.1.0,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry]
name = "udemy-tensorflow-developer-certificate-in-2023-zero-to-mastery"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.13,<4.0"
djoser = "^2.2.0"
jupyter = "^1.0.0"
scikit-learn = "1.2.2"
